+++ RealmServer/PacketServer/SMSG_CHAR_CREATE.cs
@@ -6,7 +6,7 @@ namespace RealmServer.PacketServer
     {
         public SMSG_CHAR_CREATE(LoginErrorCode code) : base(RealmEnums.SMSG_CHAR_CREATE)
         {
-            Write((byte)code);
+            Write((byte) code);
         }
     }
 }