+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.6] - 2018-10-09
+### Added
+- Common.Globals.Enum 
+  - CharacterFlag
+- RealmEnums.SMSG_ADDON_INFO Partial
+- RealmEnums.SMSG_CHAR_ENUM
+- RealmServer Database
+  - GetCharacters (from user)
+  - FindCharacaterByName
+
+### Changed
+- OnAuthSession documented.
+
 ## [0.0.5] - 2018-10-09
 ### Added
 - RealmEnums.CMSG_PING